Removed correlation feature importance
Due to time constraints, I have decided to remove the feature importance check. It's a nice robustness check for the correlation analysis, but the implementation was flawed as it didn't output what I thought it did. Trying to fix it will cost more time than it's currently worth, but I'll re-implement it if I have enough time later.